Generic customSWR and ScheduleTab abstracted
